backend/database.py
import os
from dotenv import load_dotenv
from supabase import create_client, Client

# Load env from backend/.env before reading variables
load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")

if not SUPABASE_URL or not SUPABASE_KEY:
    raise RuntimeError("SUPABASE configuration missing. Set SUPABASE_URL to https://<project>.supabase.co and SUPABASE_KEY.")

supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)

def get_supabase() -> Client:
    return supabase