backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pydantic
httpx
python-jose[cryptography]
passlib[bcrypt]
supabase==2.18.1
bcrypt